requirements.txt
openai
gradio
langchain
python-dotenv
black
isort